Fix URL encoding for query values.
Same as previous fix:
https://github.com/chmike/CxxUrl/commit/614515e12a762715ee671a4d92692c513abdca85

+++ url.cpp
@@ -514,7 +514,7 @@ class encode_query_val {
             else if (is_char(c,e.m_mask))
                 o<<c;
             else
-                o<<'%'<<"0123456789ABCDEF"[c>>4]<<"0123456789ABCDEF"[c&0xF];
+                o<<'%'<<"0123456789ABCDEF"[((uint8_t)c)>>4]<<"0123456789ABCDEF"[((uint8_t)c)&0xF];
         return o;
     }
 };
